Add Theme Applier for plugin
This was necessary, because usually Pascal ABC.NET IDE is inside Program Files. To modify it, the user should run IDE with Admin rights. Instead of asking a user to run always as admin, I thought it'd be better to run another app and by that, the user will be asked to give admin rights to the app.